+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 1.12.5 - 2024-03-xx
+## Version 1.12.5 - 2024-02-09
 
 ### Fixed
 
-- tbd
+- Make data type conversion of values more robust
+- Fix bound action call error, where action parameter name also is an entity element name
 
 ## Version 1.12.4 - 2024-02-09